refactor(analytics): split report queries into modules
